added nullable decimal, fixed some errant sql emitters recently broken

+++ FlitBit.Data/SqlServer/SqlMappedNullableDoubleEmitter.cs
@@ -6,7 +6,8 @@ using FlitBit.Emit;
 
 namespace FlitBit.Data.SqlServer
 {
-  internal class SqlMappedNullableDoubleEmitter : SqlDbTypeNullableEmitter<double>
+
+  internal class SqlMappedNullableDoubleEmitter : SqlDbTypeNullableEmitter<decimal>
   {
     internal SqlMappedNullableDoubleEmitter()
       : base(DbType.Double, SqlDbType.Float)
